+++ templates/python.flask/{{cookiecutter.project_safe_name}}/ci/analysis.sh
@@ -10,6 +10,11 @@ fi
 chmod +x ./venv/bin/activate
 
 ./venv/bin/activate
-pip install --trusted-host scmesos06 -i http://scmesos06/simple -r requirements_dev.txt --cache-dir=/tmp/recipe
 
-${PYLINT} -f parseable {{cookiecutter.project_slug}} | tee pylint.out
+pip install --trusted-host scmesos06 -i http://scmesos06/simple -r requirements_dev.txt --cache-dir=/tmp/recipe
+[ $? -gt 0 ] && exit 1
+
+pip install --trusted-host scmesos06 -i http://scmesos06/simple Coverage
+[ $? -gt 0 ] && exit 1
+
+${PYLINT} -f parseable app | tee pylint.out