+++ templates/python.flask/{{cookiecutter.project_safe_name}}/ci/at.sh
@@ -2,6 +2,16 @@
 
 IMAGE=$(sirius docker_image_name | head -n 1)
 
+# ----------------------------------------
+# deploy app manually
 # replace DEPLOY_SERVER with the release server
+#
+# sirius docker_deploy:{{cookiecutter.project_slug}},${IMAGE},server=DEPLOY_SERVER,ports="9201;8080"
+# -----------------------------------------
 
-sirius docker_deploy:{{cookiecutter.project_slug}},${IMAGE},server=DEPLOY_SERVER,ports="9201;8080"
+# ----------------------------------------
+# deploy app into GDEV apps cluster
+# you should not care about the app's tcp port
+# ---------------------------------------
+
+sirius docker_dev_deploy:{{cookiecutter.project_slug}},${IMAGE}