+++ templates/python.lib/{{cookiecutter.project_safe_name}}/ci/ut.sh
@@ -1,6 +1,8 @@
 #! /bin/sh
 
-COVER=coverage
+COVER=./venv/bin/coverage
+
+./venv/bin/activate
 
 ${COVER} run --source {{cookiecutter.project_slug}} -m unittest discover --start-directory test --pattern test_*.py